Added tilde expansion, edited variable expansion
The tilde now expands to the user's home directory, if it is at the start of a word/dquote.
For some reason, the tilde gets included with the quote even if it is outside.
This could be a potential bug, keeping an eye out.

Variable expansion was also modified to match on a string and iteratively modify it,
instead of matching on elements of a vec.