Added matching on tilde in initial parse
Tilde detection now happens in the initial parse loop. When the parser encounters a tilde and the previous character is a space, it enters an in_tilde mode that terminates when:
- It encounters a metacharacter (including spaces)
- The user enters quote mode (single or double) while in_tilde and then exits quote mode: This also exits in_tilde mode.

Adding tilde detection here should speed things up: running the expand_tilde for every word in the sequence takes cpu cycles,
but also it allows us to escape ~ with a backslash. The escape mechanism works at the initial parse: when the parse loop encounters a backslash;
this triggers escape mode that blocks all metacharacters and forces the character to be pushed to the buffer, and then ends after this character.

Moving tilde detection to the initial parse loop allows backslashes to cancel the special meaning of tildes.

Todo: Do this for variables ($abc) as well.