Implemented quote tracking for backquotes
Previously, the shell took the input and matched on a regex, then parsed and executed the match like any command, replacing all the text in that match with the command output.
However, this prevented the shell from tracking the bquotes, and prompting the user to complete their command.
Thus, bquote support is now implemented in the tokenizer directly, and is able to act with respect to the context. It also knows when it is in a dquote containing a bquote,
and when the dquote ends  will push a bquote token instead.
The parser will now only execute command expansion for bquote tokens. It still uses the regex to match the parts of the command enclosed within backquotes,
as the bquote token may contain more than the commands.

Other changes included changing the function signature of the tokenize function, since the shell is no longer needed, hence the widespread changes.

Right now, backslash escape is not supported for backquotes, but I hope to implement this eventually.