Variable expansion now occurs at the execute stage. This is to accomodate variable expansion in scripting, as the full syntax is only parsed during the execution stage after tokenization and parsing, which was where expansion occurred beforehand.
Jobs now carry a new type TokenCmd, which encodes the quotation type of the string.
The impl of Cmd contains an assoc fn from_tokencmd() which does the conversion to Cmd that the main execute function accepts.
Variable expansion occurs in this function, and this function is called in execute_jobs(), which is the entry point for scripting.

The refactoring might still be buggy and edge cases are most definitely not covered yet.