Implemented for loop range expansion
For loops parsing can now expand ranges: [1..10], and will return errors if the range cannot be parsed into integers.
The range is expanded from the first number to just before the second number.
To support range detection (and also if statement equality checking), the sqbrkt token was added to the tokenizer.
This is ignored by the from_tokencmd function, so it can be detected only by the scripting parser and expanded accordingly
and is otherwise treated as a normal word by the shell.

Also included in this commit is some (not working) code for ! evaluation for the if statement because I forgot to git stash),
and additional documentation in types.