Implemented file globbing
The shell will now determine whether a pattern needs globbing and expand it accordingly.
For a pattern to be marked as needing globbing, it needs to contain at least one * or one /.
These characters cannot be escaped, and to null out their effect they need to be enclosed in quotes.